correction erreurs minimess

+++ Simulator/src/fastfood/Burger.java
@@ -6,7 +6,7 @@ import modele.Objet;
 public class Burger extends Objet 
 {
 
-	public Burger(Case etat, double timeout) 
+	public Burger(Case etat, long timeout) 
 	{
 		super(etat, timeout, 1000) ;
 	}
