+++ Simulator/src/fastfood/Client.java
@@ -6,7 +6,7 @@ import modele.Objet;
 public class Client extends Objet
 {
 	
-	public Client(Case etat, double timeout,  int nombremax) 
+	public Client(Case etat, long timeout,  int nombremax) 
 	{
 		super(etat, timeout,  nombremax) ;
 	}
